fix: add keyboard shortcut (ctrl+R) to run the selected run configuration
